+++ plugins/devtools/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @backstage/plugin-devtools
 
+## 0.1.30-next.2
+
+### Patch Changes
+
+- e4ddf22: Internal update to align with new blueprint parameter naming in the new frontend system.
+- Updated dependencies
+  - @backstage/frontend-plugin-api@0.11.0-next.1
+  - @backstage/core-compat-api@0.5.0-next.2
+  - @backstage/core-components@0.17.5-next.1
+  - @backstage/core-plugin-api@1.10.9
+  - @backstage/errors@1.2.7
+  - @backstage/plugin-devtools-common@0.1.17
+  - @backstage/plugin-permission-react@0.4.36
+
 ## 0.1.30-next.1
 
 ### Patch Changes